# Remove `_full_jacobian` from base stepper