--- a/core/include/detray/propagator/actors/parameter_transporter.hpp
+++ b/core/include/detray/propagator/actors/parameter_transporter.hpp
@@ -19,31 +19,14 @@ namespace detray {
 template <typename algebra_t>
 struct parameter_transporter : actor {
 
+    // Bound matrix type
+    using bound_matrix_t = bound_matrix<algebra_t>;
+    // Matrix actor
+    using matrix_operator = dmatrix_operator<algebra_t>;
+
     struct state {};
 
-    /// bound to free jacobian from previous surface
-    struct bound_to_free_jacobian_kernel {
-
-        // Transformation matching this struct
-        using transform3_type = dtransform3D<algebra_t>;
-
-        template <typename mask_group_t, typename index_t,
-                  typename propagator_state_t>
-        DETRAY_HOST_DEVICE inline bound_to_free_matrix<algebra_t> operator()(
-            const mask_group_t& mask_group, const index_t& index,
-            const transform3_type& trf3, propagator_state_t& propagation) {
-
-            using frame_t = typename mask_group_t::value_type::shape::
-                template local_frame_type<algebra_t>;
-
-            using jacobian_engine_t = detail::jacobian_engine<frame_t>;
-
-            return jacobian_engine_t::bound_to_free_jacobian(
-                trf3, mask_group[index], propagation._stepping._bound_params);
-        }
-    };
-
-    struct kernel {
+    struct full_jacobian_kernel {
 
         /// @name Type definitions for the struct
         /// @{
@@ -52,8 +35,6 @@ struct parameter_transporter : actor {
         using transform3_type = dtransform3D<algebra_t>;
         // scalar_type
         using scalar_type = dscalar<algebra_t>;
-        // Matrix actor
-        using matrix_operator = dmatrix_operator<algebra_t>;
         // 2D matrix type
         template <std::size_t ROWS, std::size_t COLS>
         using matrix_type = dmatrix<algebra_t, ROWS, COLS>;
@@ -62,7 +43,7 @@ struct parameter_transporter : actor {
 
         template <typename mask_group_t, typename index_t,
                   typename propagator_state_t>
-        DETRAY_HOST_DEVICE inline void operator()(
+        DETRAY_HOST_DEVICE inline bound_matrix_t operator()(
             const mask_group_t& /*mask_group*/, const index_t& /*index*/,
             const transform3_type& trf3,
             const bound_to_free_matrix<algebra_t>& bound_to_free_jacobian,
@@ -73,8 +54,6 @@ struct parameter_transporter : actor {
 
             using jacobian_engine_t = detail::jacobian_engine<frame_t>;
 
-            using bound_matrix_t = bound_matrix<algebra_t>;
-
             using free_matrix_t = free_matrix<algebra_t>;
             using free_to_bound_matrix_t =
                 typename jacobian_engine_t::free_to_bound_matrix_type;
@@ -84,10 +63,6 @@ struct parameter_transporter : actor {
 
             // Free vector
             const auto& free_params = stepping();
-
-            // Convert free to bound vector
-            stepping._bound_params.set_parameter_vector(
-                detail::free_to_bound_vector<frame_t>(trf3, free_params));
 
             // Free to bound jacobian at the destination surface
             const free_to_bound_matrix_t free_to_bound_jacobian =
@@ -106,32 +81,17 @@ struct parameter_transporter : actor {
                     .template identity<e_free_size, e_free_size>() +
                 path_correction;
 
-            stepping._full_jacobian = free_to_bound_jacobian * correction_term *
-                                      free_transport_jacobian *
-                                      bound_to_free_jacobian;
-
-            bound_matrix_t new_cov =
-                stepping._full_jacobian * stepping._bound_params.covariance() *
-                matrix_operator().transpose(stepping._full_jacobian);
-
-            // Calculate surface-to-surface covariance transport
-            stepping._bound_params.set_covariance(new_cov);
+            return free_to_bound_jacobian * correction_term *
+                   free_transport_jacobian * bound_to_free_jacobian;
         }
     };
 
     template <typename propagator_state_t>
-    DETRAY_HOST_DEVICE void operator()(state& /*actor_state*/,
-                                       propagator_state_t& propagation) const {
-        auto& stepping = propagation._stepping;
+    DETRAY_HOST_DEVICE inline bound_matrix_t get_full_jacobian(
+        propagator_state_t& propagation) const {
+        const auto& stepping = propagation._stepping;
         const auto& navigation = propagation._navigation;
 
-        // Do covariance transport when the track is on surface
-        if (!(navigation.is_on_sensitive() ||
-              navigation.encountered_sf_material())) {
-            return;
-        }
-
-        using matrix_operator = dmatrix_operator<algebra_t>;
         using detector_type = typename propagator_state_t::detector_type;
         using geo_cxt_t = typename detector_type::geometry_context;
         const geo_cxt_t ctx{};
@@ -148,16 +108,50 @@ struct parameter_transporter : actor {
             tracking_surface<detector_type> prev_sf{navigation.detector(),
                                                     stepping._prev_sf_id};
 
-            bound_to_free_jacobian =
-                prev_sf.template visit_mask<bound_to_free_jacobian_kernel>(
-                    prev_sf.transform(ctx), propagation);
+            bound_to_free_jacobian = prev_sf.bound_to_free_jacobian(
+                ctx, propagation._stepping._bound_params);
         }
 
-        sf.template visit_mask<kernel>(sf.transform(ctx),
-                                       bound_to_free_jacobian, propagation);
+        return sf.template visit_mask<full_jacobian_kernel>(
+            sf.transform(ctx), bound_to_free_jacobian, propagation);
+    }
+
+    template <typename propagator_state_t>
+    DETRAY_HOST_DEVICE void operator()(state& /*actor_state*/,
+                                       propagator_state_t& propagation) const {
+
+        using detector_type = typename propagator_state_t::detector_type;
+        using geo_cxt_t = typename detector_type::geometry_context;
+        const geo_cxt_t ctx{};
+
+        auto& stepping = propagation._stepping;
+        const auto& navigation = propagation._navigation;
+
+        // Do covariance transport when the track is on surface
+        if (!(navigation.is_on_sensitive() ||
+              navigation.encountered_sf_material())) {
+            return;
+        }
+
+        const bound_matrix_t full_jacobian = get_full_jacobian(propagation);
+
+        // Calculate surface-to-surface covariance transport
+        const bound_matrix_t new_cov =
+            full_jacobian * stepping._bound_params.covariance() *
+            matrix_operator().transpose(full_jacobian);
+
+        // Current surface
+        const auto sf = navigation.get_surface();
+
+        // Convert free to bound vector
+        stepping._bound_params.set_parameter_vector(
+            sf.free_to_bound_vector(ctx, stepping()));
+
+        stepping._bound_params.set_covariance(new_cov);
 
         // Set surface link
-        stepping._bound_params.set_surface_link(sf.barcode());
+        stepping._bound_params.set_surface_link(
+            navigation.get_surface().barcode());
     }
 };  // namespace detray